requirements.txt
# Core
langchain==0.1.0
openai==1.0.0
google-generativeai==0.3.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# UI
streamlit==1.29.0
pillow==10.1.0

# Publishing
tweepy==4.14.0
python-linkedin-v2==0.1.0
facebook-sdk==3.1.0

# Scheduling
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.12.1
flake8==7.0.0
mypy==1.7.1
